requirements.txt
numpy
biopython==1.73
